Organize parsing utilities
