+++ src/app_controller.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional
 
 from character_role_suggester import CharacterRoleSuggester
-from markdown_document import MarkdownDocument
+from parsing.markdown_document import MarkdownDocument
 
 logger = logging.getLogger(__name__)
