+++ src/batch_processing/base_batch_processor.py
@@ -4,7 +4,7 @@ from typing import Dict, List, Optional, Set, Union
 
 from models.document_model import H3Pydantic, PanelPydantic
 from logging_config import get_logger
-from markdown_document import MarkdownDocument
+from parsing.markdown_document import MarkdownDocument
 from models.section_titles import SECTION_TITLES, normalize_section_name
 
 logger = get_logger(__name__)
