+++ src/batch_processing/batch_processor.py
@@ -6,7 +6,7 @@ import openai_service
 from batch_processing.base_batch_processor import BaseBatchProcessor
 from models.document_model import H3Pydantic, PanelPydantic
 from logging_config import get_logger
-from markdown_document import MarkdownDocument
+from parsing.markdown_document import MarkdownDocument
 
 logger = get_logger(__name__)
