+++ src/character_role_suggester.py
@@ -3,7 +3,7 @@
 from pathlib import Path
 from typing import Dict, List
 
-from markdown_document import MarkdownDocument
+from parsing.markdown_document import MarkdownDocument
 from openai_service import suggest_character_roles_for_panels
 from models.section_titles import SECTION_TITLES
