+++ src/parsing/panel_section_manager.py
