+++ src/role_validator_tool.py
@@ -10,7 +10,7 @@ from typing import Any, Dict, List, Set, Union
 
 from models.document_model import PanelPydantic
 from logging_config import get_logger
-from markdown_document import MarkdownDocument
+from parsing.markdown_document import MarkdownDocument
 from openai_service import suggest_character_roles_from_context
 from models.section_titles import SECTION_TITLES
